refactor(Map): Put the Map component into the Map Folder
